Use `electrode-hybrid-fetch` to handle SSR pre-fetching
We keep a reference to the HAPI.js server instance and pass it to `electrode-hybrid-fetch`.
We then use the `redux-thunk` middleware and configure it to supply an extra argument to our function actions when dispatched. We set this argument to be either the `ssrFetch` of `electrode-hybrid-fetch` (in the server side store creation) or the regular `fetch` for use in client side store creation.